gen/requirements.txt
boto3==1.14.49
botocore==1.17.49
jmespath==0.10.0
pypng==0.0.20
PyQRCode==1.2.1
s3transfer==0.3.3
six==1.15.0
Werkzeug==1.0.1
